requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
llama-api-client==0.1.1
moviepy==2.2.1
numpy<=1.26
openai-whisper
opencv-python==4.11.0.86
pillow==11.2.1
proglog==0.1.12
pydantic==2.11.5
pydantic-core==2.33.2
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.1
urllib3==2.4.0
whisperx